Refining code structure
* Move types to appropriate modules and remove src/types file
* Remove borders constants exports
* Update tags, badges and description
* Remove sourceMap generating since output files are not minified and clear enough
* Up library version to 1.1.5
