+++ README.md
@@ -1,49 +1,55 @@
 # hydrate-text
-Tiny library for dynamic text hydrating
+Tiny library for dynamic text hydrating with variables
 
 [![Version][version-badge]][package]
 [![MIT License][license-badge]][license]
+![Downloads][downloads-badge]
 [![GitHub Release][release-badge]][release]
+[![Build Status][build-badge]][build]
 
 [![Minified size][min-size-badge]][size]
 [![Minified and gzipped size][minzip-size-badge]][size]
 [![Dependencies Status][dependencies-badge]][dependencies]
 [![devDependencies Status][dev-dependencies-badge]][dev-dependencies]
 
-[![Build Status][build-badge]][build]
+![Types][types-badge]
 [![Code Coverage][coverage-badge]][coverage]
 [![Maintainability][maintainability-badge]][maintainability]
 [![Known Vulnerabilities][vulnerabilities-badge]][vulnerabilities]
 
-[version-badge]: https://img.shields.io/npm/v/hydrate-text.svg
+[version-badge]: https://badgen.net/npm/v/hydrate-text?style=flat
 [package]: https://www.npmjs.com/package/hydrate-text
 
-[release-badge]: https://img.shields.io/github/v/release/vasilii-kovalev/hydrate-text.svg
-[release]: https://github.com/vasilii-kovalev/hydrate-text/releases
+[downloads-badge]: https://badgen.net/npm/dt/hydrate-text?style=flat&color=blue
 
-[license-badge]: https://img.shields.io/npm/l/hydrate-text.svg
+[license-badge]: https://badgen.net/github/license/vasilii-kovalev/hydrate-text?style=flat
 [license]: https://github.com/vasilii-kovalev/hydrate-text/blob/master/LICENSE
 
-[min-size-badge]: https://badgen.net/bundlephobia/min/hydrate-text
-[minzip-size-badge]: https://badgen.net/bundlephobia/minzip/hydrate-text
-[size]: https://bundlephobia.com/result?p=hydrate-text@latest
+[release-badge]: https://badgen.net/github/release/vasilii-kovalev/hydrate-text?style=flat
+[release]: https://github.com/vasilii-kovalev/hydrate-text/releases
 
-[dependencies-badge]: https://david-dm.org/vasilii-kovalev/hydrate-text.svg
-[dependencies]: https://david-dm.org/vasilii-kovalev/hydrate-text
-[dev-dependencies-badge]: https://david-dm.org/vasilii-kovalev/hydrate-text/dev-status.svg
-[dev-dependencies]: https://david-dm.org/vasilii-kovalev/hydrate-text?type=dev
-
-[build-badge]: https://github.com/vasilii-kovalev/hydrate-text/workflows/build-test/badge.svg?branch=master
+[build-badge]: https://badgen.net/github/status/vasilii-kovalev/hydrate-text?style=flat
 [build]: https://github.com/vasilii-kovalev/hydrate-text/actions?query=workflow%3Abuild-test+branch%3Amaster
 
-[maintainability-badge]: https://api.codeclimate.com/v1/badges/26bb75b93d63c800e6ae/maintainability
+[min-size-badge]: https://badgen.net/bundlephobia/min/hydrate-text?style=flat
+[minzip-size-badge]: https://badgen.net/bundlephobia/minzip/hydrate-text?style=flat
+[size]: https://bundlephobia.com/result?p=hydrate-text
+
+[dependencies-badge]: https://badgen.net/david/dep/vasilii-kovalev/hydrate-text?style=flat
+[dependencies]: https://david-dm.org/vasilii-kovalev/hydrate-text
+[dev-dependencies-badge]: https://badgen.net/david/dev/vasilii-kovalev/hydrate-text?style=flat
+[dev-dependencies]: https://david-dm.org/vasilii-kovalev/hydrate-text?type=dev
+
+[types-badge]: https://badgen.net/npm/types/hydrate-text?style=flat
+
+[maintainability-badge]: https://badgen.net/codeclimate/maintainability/vasilii-kovalev/hydrate-text?style=flat
 [maintainability]: https://codeclimate.com/github/vasilii-kovalev/hydrate-text/maintainability
 
-[coverage-badge]: https://coveralls.io/repos/github/vasilii-kovalev/hydrate-text/badge.svg?branch=master
-[coverage]: https://coveralls.io/github/vasilii-kovalev/hydrate-text?branch=master
+[coverage-badge]: https://badgen.net/coveralls/c/github/vasilii-kovalev/hydrate-text?style=flat
+[coverage]: https://coveralls.io/github/vasilii-kovalev/hydrate-text
 
-[vulnerabilities-badge]: https://snyk.io/test/github/vasilii-kovalev/hydrate-text/badge.svg?targetFile=package.json
-[vulnerabilities]: https://snyk.io/test/github/vasilii-kovalev/hydrate-text?targetFile=package.json
+[vulnerabilities-badge]: https://badgen.net/snyk/vasilii-kovalev/hydrate-text?style=flat
+[vulnerabilities]: https://snyk.io/test/github/vasilii-kovalev/hydrate-text
 
 ## Installation
 ```shell
@@ -51,10 +57,9 @@ npm install hydrate-text
 ```
 
 ## Features
-* Dependency-free
-* Side-effects-free
-* Tree-shakable
 * Light-weight
+* Dependency-free
+* Tree-shakable
 * Works with variables as an object and an array
 * Flexible variable syntax change
 * Strongly typed with TypeScript
@@ -75,8 +80,11 @@ console.log(hydrateText(text2, ['run tests', 'make tea']));
 
 // '/users/1'
 console.log(hydrateText(route, { userId: 1 }, { start: ':' }));
+```
 
-/* Can be configured */
+Initial variables syntax can be configured via `configureHydrateText`, that
+returns `hydrateText` function as a result.
+```typescript
 import { configureHydrateText } from 'hydrate-text';
 
 const route = '/users/:userId';
