+++ package.json
@@ -1,36 +1,30 @@
 {
   "name": "hydrate-text",
-  "version": "1.1.4",
-  "description": "Tiny library for dynamic text hydrating",
+  "version": "1.1.5",
+  "description": "Tiny library for dynamic text hydrating with variables",
   "main": "lib/index",
   "module": "lib/index",
   "types": "lib/index.d.ts",
   "sideEffects": false,
-  "scripts": {
-    "test": "jest",
-    "tslint": "tsc --noEmit",
-    "eslint": "eslint --ext .ts src --color",
-    "build": "tsc",
-    "prepare": "npm run build",
-    "prepublishOnly": "npm run tslint && npm run eslint && npm test",
-    "postversion": "git push && git push --tags"
-  },
+  "keywords": [
+    "javascript",
+    "typescript",
+    "template",
+    "string",
+    "formatter",
+    "zero-dependency",
+    "tree-shakable"
+  ],
+  "author": "Vasilii Kovalev <vasya.kowalev@yandex.ru>",
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/vasilii-kovalev/hydrate-text.git"
   },
-  "keywords": [
-    "javascript",
-    "typescript",
-    "zero-dependency",
-    "text"
-  ],
-  "author": "Vasilii Kovalev <vasya.kowalev@yandex.ru>",
-  "license": "MIT",
+  "homepage": "https://github.com/vasilii-kovalev/hydrate-text#readme",
   "bugs": {
     "url": "https://github.com/vasilii-kovalev/hydrate-text/issues"
   },
-  "homepage": "https://github.com/vasilii-kovalev/hydrate-text#readme",
   "devDependencies": {
     "@types/jest": "^26.0.5",
     "@typescript-eslint/eslint-plugin": "^3.7.0",
@@ -42,6 +36,15 @@
     "jest": "^26.1.0",
     "ts-jest": "^26.1.3",
     "typescript": "^3.9.7"
+  },
+  "scripts": {
+    "test": "jest",
+    "tslint": "tsc --noEmit",
+    "eslint": "eslint --ext .ts src --color",
+    "build": "tsc",
+    "prepare": "npm run build",
+    "prepublishOnly": "npm run tslint && npm run eslint && npm test",
+    "postversion": "git push && git push --tags"
   },
   "jest": {
     "preset": "ts-jest",
