+++ src/utils.ts
@@ -1,4 +1,23 @@
-import { EscapeRegExp, IsNil } from './types';
+/**
+  * Source: https://github.com/lodash/lodash/blob/master/escapeRegExp.js
+  * Escapes the `RegExp` special characters "^", "$", "\", ".", "*", "+",
+  * "?", "(", ")", "[", "]", "{", "}", and "|" in `value`.
+  * @param {string} [value=''] The string to escape.
+  * @returns {string} Returns the escaped string.
+*/
+export interface EscapeRegExp {
+  (value?: string): string;
+}
+
+/**
+  * Source: https://github.com/lodash/lodash/blob/master/isNil.js
+  * Checks if `value` is `null` or `undefined`.
+  * @param {*} value The value to check.
+  * @returns {boolean} Returns `true` if `value` is nullish, else `false`.
+*/
+export interface IsNil {
+  (value: any): boolean;
+}
 
 /**
   * Used to match `RegExp`
