+++ tsconfig.json
@@ -16,7 +16,7 @@
     // "jsx": "preserve",
     "declaration": true,
     // "declarationMap": true,
-    "sourceMap": true,
+    // "sourceMap": true,
     // "outFile": "./",
     "outDir": "./lib",
     "rootDir": "src",
